Remove code no longer necessary, fix some minor bugs.
__init__.py:
Fix bug in help statement for "--max-1dpamzt" argument.
Replace "minus_z" with "minus_yz" in plot title geneneration code.

make_sim_inputs.py:
Remove import of Ska.DBI since it is not used anywhere
Remove import of characteristics file since it is not used as of this release.
Remove definition and references to "psmc_powers" since it is no longer used as of this release.

characteristics.py:
Remove file; this is no longer used with migration of PSMC model to Xija framework.

twodof.py:
Remove file; this is no longer used with migration of PSMC model to Xija framework.

Signed-off-by: Matthew Dahmer <matthew.dahmer@gmail.com>